Split config into config and models
